requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-core
python-dotenv
requests
wikipedia
chainlit